Remove LOC-specific stuff from user admin.
